Eledest child variables and test Co-authored-by: Max Ghenis <mghenis@gmail.com>
