+++ policyengine_canada/parameters/gov/cra/tax/income/credits/climate_action_incentive/adult_age.yaml
@@ -1,6 +1,17 @@
 description: Climate Action Incentive adult age.
 values:
-  2022-07-01: 19
+  2019-01-01:
+    value: 18
+    metadata:
+      reference:
+        - title: Income Tax Act
+          href: https://laws-lois.justice.gc.ca/eng/acts/I-3.3/20190101/P1TT3xt3.html
+  2022-07-01:
+    value: 19
+    metadata:
+      reference:
+        - title: Income Tax Act
+          href: https://laws-lois.justice.gc.ca/eng/acts/I-3.3/20220701/P1TT3xt3.html
 metadata:
   unit: year
   name: climate_action_incentive_adult_age
