+++ policyengine_canada/tests/gov/cra/tax/income/credits/climate_action/is_eldest_child_for_climate_action_incentive.yaml
@@ -1,0 +1,38 @@
+- name: Single person with no children should be false.
+  period: 2023
+  input:
+    age: 20
+  output:
+    is_eldest_child_for_climate_action_incentive: false
+
+- name: Parent with two children.
+  period: 2023
+  input:
+    people:
+      parent:
+        age: 30
+      c1:
+        age: 10
+      c2:
+        age: 9
+    household:
+      members: [parent, c1, c2]
+  output:
+    is_eldest_child_for_climate_action_incentive: [false, true, false]
+
+- name: Parent with two 18-year-olds; select the first only.
+  period: 2023
+  input:
+    people:
+      parent:
+        age: 30
+      c1:
+        age: 18
+        is_dependent: true
+      c2:
+        age: 18
+        is_dependent: true
+    household:
+      members: [parent, c1, c2]
+  output:
+    is_eldest_child_for_climate_action_incentive: [false, true, false]
