+++ policyengine_canada/variables/gov/cra/tax/income/credits/climate_action/climate_action_incentive_dependent_children.py
@@ -1,0 +1,12 @@
+from policyengine_canada.model_api import *
+
+
+class climate_action_incentive_dependent_children(Variable):
+    value_type = int
+    entity = Household
+    label = "Climate action incentive dependent Children"
+    unit = CAD
+    documentation = "Number of eligible dependent children"
+    definition_period = YEAR
+
+    formula = sum_of_variables(["is_child_for_climate_action_incentive"])
