+++ policyengine_canada/variables/gov/cra/tax/income/credits/climate_action/climate_action_incentive_single_parent.py
@@ -12,7 +12,7 @@ class climate_action_incentive_single_parent(Variable):
     def formula(household, period, parameters):
         single_parent = household("is_single_parent", period)
         province = household("province_str", period)
-        is_first_child = household("is_first_child", period)
+        is_first_child = household("is_eldest_child", period)
         first_child_amount = parameters(
             period
         ).gov.cra.tax.income.credits.climate_action_incentive.amount.first_child_in_single_parent_family[
