+++ policyengine_canada/variables/household/person/is_dependent.py
@@ -6,3 +6,7 @@ class is_dependent(Variable):
     entity = Person
     label = "Is a dependent"
     definition_period = YEAR
+
+    # Impute dependent status on age.
+    def formula(person, period, parameters):
+        return person("age", period) < 18
